Add files via upload
In this project, which is one of the most advanced projects I have done so far, Java (Object Oriented Programming Language and SQL) and MsSQL (Structured Query Language)
I made hospital automation with There may be some deficiencies in my codes if you see this and let me know, I would appreciate it. This project has been made by taking advantage of the features of object-oriented programming and this has given me a better understanding of the logic of object-oriented programming. You can also see the example of the connection between languages in this project where I use JAVA and MsSQL languages together.